backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
msal>=1.24.0
requests>=2.31.0
djangorestframework-simplejwt>=5.3.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
